activity

+++ app/src/main/java/com/example/smartcalculator/Solution/CoefficientsTransformer.java
@@ -5,7 +5,7 @@ import java.util.Map;
 public class CoefficientsTransformer {
     public static void deleteZeros(Coefficients coefficients)
     {
-        for (Map.Entry<Integer, Integer> entry : coefficients.entrySet()){
+        for (Map.Entry<Integer, Double> entry : coefficients.entrySet()){
             if (entry.getValue() == 0)
             {
                 coefficients.remove(entry.getKey());
